feat: Introduce ConsumerBackend::requestByTypeAndFilter()
The existing findTypeAndFilter() method returns a ResourceProxyIterator
object, which is meant to be fully populated when it gets returned,
either by fetching or by loading from cache.

The new requestByTypeAndFilter() method returns a promise of a
ResourceProxyIterator, which might be fulfilled in case there was
a cache entry, but might as well just be on its way.

This is meant for calling multiple remotes at the same time and
waiting for all to return, instead of calling one after the other.